% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignation.R
\name{assignation}
\alias{assignation}
\alias{assignation_en}
\alias{assignation_fr}
\title{Insert a GitHub assignation in the document}
\usage{
assignation(
  name,
  url,
  course.urls = NULL,
  part = NULL,
  course.names = c(course1 = "Data Science"),
  toc = "",
  texts = assignation_en(),
  assign.img = "images/list-assign.png",
  assign.link = "github_assignation"
)

assignation_en(title, part.name, alt, sub, course, toc.def)

assignation_fr(title, part.name, alt, sub, course, toc.def)
}
\arguments{
\item{name}{The name of the assignation (usually the same as the name of the
template GitHub repository).}

\item{url}{The URL of the assignation (could be a named list for different
versions depending on the course).}

\item{course.urls}{Named vector with the Classroom URLS for each course.
Names are the course identifiers in Moodle. If \code{NULL}, no course-specific
sections are added.}

\item{part}{If the assignation presents several parts in the \code{README.md} file
(because the same assignation serves at different places), indicate the part
here. Otherwise, leave the default value \code{NULL} if there are no parts.}

\item{course.names}{A named character vector with the name of the course.
Names used must be the same as for \code{course.urls}, but there may be more here,
and not necessarily in the same order.}

\item{toc}{The exercise table of content (ex-toc) label. If \code{toc = ""}
(default), a generic label is calculated using \code{toc.def} from \code{texts}.}

\item{texts}{Various sentences used to construct the assignation bloc. You
can make a call to \code{assignation_en()} or \code{assignation_fr()} as a basis
and modify only the sentence you want.}

\item{assign.img}{The relative path to the image to use before the label in
the ex-toc.}

\item{assign.link}{The link to the learnr help page (when the user clicks on
the image in the ex-toc).}

\item{title}{The title of the block.}

\item{part.name}{The word to use for "part".}

\item{alt}{The text to display for alternate access to the repository (for
non-registered users). Use a string following the \code{\link[=glue]{glue()}} syntax to replace
variables.}

\item{sub}{The text that appears at the bottom of the assignation block.}

\item{course}{The text for items corresponding to courses.}

\item{toc.def}{The default ex-toc label (using \code{\link[=glue]{glue()}} syntax).}
}
\value{
Markdown code that generates the GitHub assignation block. It is most
conveniently used inside and R inline expression in your R Markdown document
on its own line with one blank line above and bellow it.
}
\description{
Insert a block of class \code{assign} with data related to a GitHub (Classroom)
assignment.
}
\details{
If the URL contains several entries, names are used to create show/hide divs
according to the \code{icourse} user information.
}
