% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignment.R
\name{assignment}
\alias{assignment}
\alias{assignment2}
\alias{challenge}
\alias{challenge2}
\alias{assignment_en}
\alias{assignment_fr}
\alias{assignment2_en}
\alias{assignment2_fr}
\alias{challenge_en}
\alias{challenge_fr}
\alias{challenge2_en}
\alias{challenge2_fr}
\title{Insert a GitHub (Classroom) assignment or challenge in the document}
\usage{
assignment(
  name,
  url,
  course.ids = NULL,
  course.urls = NULL,
  course.starts = NULL,
  course.ends = NULL,
  part = NULL,
  course.names = c(course1 = "Data Science"),
  toc = "",
  clone = TRUE,
  n = 1,
  texts = assignment_en(),
  assign.img = "images/list-assign.png",
  assign.link = "github_assignment",
  block = "assign",
  template = "assignment_en.html",
  baseurl = "/"
)

assignment2(
  name,
  url,
  course.ids = NULL,
  course.urls = NULL,
  course.starts = NULL,
  course.ends = NULL,
  part = NULL,
  course.names = c(course1 = "Data Science"),
  toc = "",
  clone = TRUE,
  n = 2,
  texts = assignment2_en(),
  assign.img = "images/list-assign2.png",
  assign.link = "github_assignment",
  block = "assign2",
  template = "assignment_en.html",
  baseurl = "/"
)

challenge(
  name,
  url,
  course.ids = NULL,
  course.urls = NULL,
  course.starts = NULL,
  course.ends = NULL,
  part = NULL,
  course.names = c(course1 = "Data Science"),
  toc = "",
  clone = TRUE,
  n = 1,
  texts = challenge_en(),
  assign.img = "images/list-challenge.png",
  assign.link = "github_challenge",
  block = "challenge",
  template = "assignment_en.html",
  baseurl = "/"
)

challenge2(
  name,
  url,
  course.ids = NULL,
  course.urls = NULL,
  course.starts = NULL,
  course.ends = NULL,
  part = NULL,
  course.names = c(course1 = "Data Science"),
  toc = "",
  clone = TRUE,
  n = 2,
  texts = challenge2_en(),
  assign.img = "images/list-challenge2.png",
  assign.link = "github_challenge",
  block = "challenge2",
  template = "assignment_en.html",
  baseurl = "/"
)

assignment_en(title, part.name, alt, sub, course, toc.def)

assignment_fr(title, part.name, alt, sub, course, toc.def)

assignment2_en(title, part.name, alt, sub, course, toc.def)

assignment2_fr(title, part.name, alt, sub, course, toc.def)

challenge_en(title, part.name, alt, sub, course, toc.def)

challenge_fr(title, part.name, alt, sub, course, toc.def)

challenge2_en(title, part.name, alt, sub, course, toc.def)

challenge2_fr(title, part.name, alt, sub, course, toc.def)
}
\arguments{
\item{name}{The name of the assignment or the challenge (usually the same as
the name as the GitHub Classroom item).}

\item{url}{The URL of the assignment or challenge (could be a named list for
different courses).}

\item{course.ids}{Named vector with the Classroom identifiers for the
assignments for each course and also possibly, for each group.}

\item{course.urls}{Named vector with the Classroom URLS for each course, and
also possibly for groups.
Names are the course identifiers in Moodle, or the groups defined for the
users. If \code{NULL}, no course-specific sections are added.}

\item{course.starts}{Named vector (same logic as for \verb{course.urls=}) with
starting dates (characters like "YYYY-mm-dd HH:MM:SS"). A corresponding entry
in \verb{course.urls=} is required, or the dates will be ignored.}

\item{course.ends}{Named vector (same logic as for \verb{course.urls=}) with
ending dates (characters like "YYYY-mm-dd HH:MM:SS"). A corresponding entry
in \verb{course.urls=} is required, or the dates will be ignored.}

\item{part}{If the assignment presents several parts in the \code{README.md} file
(because the same assignment is used at different places), indicate the part
here. Otherwise, leave the default value \code{NULL} if there are no parts.}

\item{course.names}{A named character vector with the name of the course.
Names used must be the same as for \code{course.urls}, but there may be more here,
and not necessarily in the same order.}

\item{toc}{The exercise table of content (ex-toc) label. If \code{toc = ""}
(default), a generic label is calculated using \code{toc.def} from \code{texts}. To not
display the exercise in the table of content, use \code{toc = NULL}.}

\item{clone}{Should the exercise be listed for cloning the repositories
(\code{TRUE} by default)? If \code{TRUE}, an entry is added in the list of assignments
whose repositories should be cloned.}

\item{n}{The number of students per project (by default, n = 1 for individual
assignments and 2 for group assignments).}

\item{texts}{Various sentences used to construct the assignment bloc. You
can make a call to \code{assignment_en()} or \code{assignment_fr()} as a basis
and modify only the sentences you want.}

\item{assign.img}{The relative path to the image to use before the label in
the ex-toc.}

\item{assign.link}{The link to the assignment help page (when the user clicks on
the image in the ex-toc).}

\item{block}{The class of the div, or the LaTeX environment to use for the
assignment block.}

\item{template}{The template file to use for the URL redirection page
(currently only assignment_en.html or assignment_fr.html).}

\item{baseurl}{The base URL for the web site.}

\item{title}{The title of the block.}

\item{part.name}{The word to use for "part".}

\item{alt}{The text to display for alternate access to the repository (for
non-registered users). Use a string following the \code{\link[=glue]{glue()}} syntax to replace
variables.}

\item{sub}{The text that appears at the bottom of the assignment block.}

\item{course}{The text for items corresponding to courses.}

\item{toc.def}{The default ex-toc label (using \code{\link[=glue]{glue()}} syntax).}
}
\value{
Markdown code that generates the GitHub Classroom assignment block.
It is most conveniently used inside an R chunk in your R Markdown document.
If you do not want to break your code chunks inside RStudio, you may use
something like \code{if (exists("assignment")) assignment(...)}.
}
\description{
Insert a block of class \code{assign}, \code{assign2}, \code{challenge} or \code{challenge2} with
data related to a GitHub (Classroom) assignment.
}
\details{
If the URL contains several entries, names are used to create show/hide divs
according to the \code{icourse} user information.
}
