% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learnitdown_learnr.R
\name{checker_ack_learnr}
\alias{checker_ack_learnr}
\title{A default checker that just acknowledges submission}
\usage{
checker_ack_learnr(
  label,
  user_code,
  solution_code,
  check_code,
  envir_result,
  evaluate_result,
  ...
)
}
\arguments{
\item{label}{The label for the learnr exercise.}

\item{user_code}{The code submitted by the user.}

\item{solution_code}{The code provided by the "-solution" chunk.}

\item{check_code}{The code provided by the "-check" chunk.}

\item{envir_result}{The environment after the execution of the chunk.}

\item{evaluate_result}{Result from evaluation of the code.}

\item{...}{Additional parameters (currently not used).}
}
\value{
A list with components \code{message}, \code{correct} and \code{location}.
}
\description{
Check code submitted during an exercise. This version just
acknowledges reception of the submission. This function is used internally
by the tutorials and is not intended for the end-user.
}
\details{
This is a simple checker function for learnitdown learnr applications that
just indicates to the user that its answer is taken into account.
}
\seealso{
\code{\link[=record_learnr]{record_learnr()}}
}
\concept{record events from the BioDataScience package}
\keyword{utilities}
