% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnose_login.R
\name{diagnose_login}
\alias{diagnose_login}
\title{Insert code in an R Markdown document to diagnose personal data information}
\usage{
diagnose_login(block = "info", lang = "en")
}
\arguments{
\item{block}{Should the diagnostic be included in a block construct
(not if \code{NULL})?}

\item{lang}{The language to use for diagnostic output (only \code{'en'} or \code{'fr'}
currently)}
}
\value{
Nothing, the function is used for its side-effect of adding
diagnostic code in the document
}
\description{
This function is used, inside an R Markdown document, to dynamically
elaborate a short report on the client-side (by using JavaScript code) that
displays the login status of the user (login, contextual data, ...).
}
\details{
Wherever you want to display this diagnostic report in your bookdown page,
add on its own line the following instruction:
'r learnitdown::diagnose_login()' between backquotes (you can also change
\code{lang}, if you want).
}
