% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5p.R
\name{h5p}
\alias{h5p}
\title{Insert H5P content in the document}
\usage{
h5p(
  id,
  baseurl,
  idurl = "wp-admin/admin-ajax.php?action=h5p_embed&id=",
  width = 780,
  height = 500,
  toc = "",
  toc.def = "H5P exercise {id}",
  h5p.img = "images/list-h5p.png",
  h5p.link = "h5p"
)
}
\arguments{
\item{id}{The ID of the H5P content in your Wordpress.}

\item{baseurl}{The first part of the URL for your domain, usually something
like \verb{https://my.site.com} \strong{without} the trailing \code{/}.}

\item{idurl}{The URL to the H5P content, usually something like \code{""}
for exercises from h5p.org or h5p.com, or
\code{"wp-admin/admin-ajax.php?action=h5p_embed&id="} for Wordpress
(by default).}

\item{width}{The width of the iframe where the H5P content is displayed.}

\item{height}{The height of the iframe.}

\item{toc}{Entry to use in the exercises table of content (\code{NULL} if no
entry, \code{""} for a default entry based on \verb{toc.def =}).}

\item{toc.def}{Text for a default toc entry using \code{\link[=glue]{glue()}} syntax for
replacement, e.g., \code{{var}}.}

\item{h5p.img}{The image to display in front of the toc entry}

\item{h5p.link}{The link when the image is clicked (sends to an help page
about learnr tutorials).}
}
\value{
HTML code that generates the iframe. It is most conveniently used
inside and R inline expression in your R Markdown document on its own line
with one blank line above and bellow it.
}
\description{
Insert H5P content in the document
}
\details{
This function is designed to work inside a Wordpress site where the H5P
plugin has been installed. You should also serve your bookdown pages as a
subdirectory inside of the same Wordpress site to allow free communication
between the parent (the bookdown page) and the child document in the iframe
(the H5P content).
}
