% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_shiny.R
\name{launch_shiny}
\alias{launch_shiny}
\title{Placeholder with screenshot to launch Shiny application on click.}
\usage{
launch_shiny(
  name,
  url,
  img,
  width = 780,
  height = 500,
  alt = "*Click to start the application.*"
)
}
\arguments{
\item{name}{Name of the shiny application (cannot be duplicated on a page).}

\item{url}{The URL of the Shiny application.}

\item{img}{The relative or absolute path to the image. Best effect is
obtained with a slightly darkened screenshot of the actual application saved
in a PNG image of exactly the same size as the iframe that embeds the Shiny
application. You can also add an icon on top of the image that suggests to
click on it to launch the application for an even better effect.}

\item{width}{The width of the image and iframe for the app.}

\item{height}{The height of image and iframe.}

\item{alt}{Alternate text to display at the bottom of the screenshot. If
\code{NULL}, nothing is displayed below the screenshot.}
}
\value{
The HTML content that creates the image and the iframe. The function
must be called from within an R inline expression of an R chunk with
\code{results='asis'} in an HTML-rendered version of the R Markdown document to
get the correct result.
}
\description{
Shiny applications can be embedded in certain R Markdown documents. However,
the application is automatically loaded at the same time as the main page,
and that may not be the desired behavior. With \code{\link[=launch_shiny]{launch_shiny()}}, you display
just a screenshot of the Shiny application in the page. The user has to click
on it to actually launch the application (that replaces the screenshot on the
page).
}
\examples{
# TODO...
}
