% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learndown-package.R
\docType{package}
\name{learndown-package}
\alias{learndown}
\alias{learndown-package}
\title{learndown: 'SciViews' - R Markdown and Bookdown Additions for Learning Material}
\description{
Extension to R Markdown and Bookdown for building better learning
  and e-learning material: H5P integration, course-contextual divs, different
  loading of Shiny and learnr applications, and much more ...
}
\details{
The SciViews learndown package provides additional R Markdown tags to build
richer learning material with interactivity.

See \code{\link[=h5p]{h5p()}} or \code{\link[=launch_shiny]{launch_shiny()}}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SciViews/learndown}
  \item \url{https://www.sciviews.org/learndown/}
  \item Report bugs at \url{https://github.com/SciViews/learndown/issues}
}

}
\author{
\strong{Maintainer}: Philippe Grosjean \email{phgrosjean@sciviews.org} (\href{https://orcid.org/0000-0002-2694-9471}{ORCID})

}
\keyword{internal}
