% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learndown_init.R
\name{learndown_init}
\alias{learndown_init}
\title{Initialize learndown features in a R MArkdown document}
\usage{
learndown_init(
  shiny = TRUE,
  h5p = TRUE,
  use.query = FALSE,
  iframe.links = TRUE,
  details.css = TRUE,
  institutions = c("institution1", "institution2"),
  courses = c("course1", "course2", "course3")
)
}
\arguments{
\item{shiny}{Do we use Shiny applications and do we want to pass parameters
and or launch the application on a click?}

\item{h5p}{Do we use H5P served from a Wordpress site in the same domain as
our R Markdown document?}

\item{use.query}{Do we collect user/course/institution data through the URL
query string (the part after the question mark in thje URL).}

\item{iframe.links}{If our document is displayed in an iframe, external link
should better tartget ther parent window. With this option, external links
with no defined target are automatically retargetted when the page loads.}

\item{details.css}{Do we want to enhance the \verb{<details>} section with a
summary surrounded by a light gray box in order to better evidence it.}

\item{institutions}{The list of possible institutions that have specific
sectionsq in this document.}

\item{courses}{The list of courses with specific sections in this document.}
}
\value{
The containt to be inserted in trhe HTML document with CSS and
Javascript code required to implement the various learndown features.
}
\description{
This function must be called at the begining of a R Markdown document (e.g,
bookdown) to installe required CSS and Javascript.
}
\examples{
# This is better placed in a setup R chunk or an R inline expression on its
# own line
learndown_init()
}
