% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learndown_learnr.R
\name{learndownLearnrSetup}
\alias{learndownLearnrSetup}
\alias{learndownLearnrBanner}
\title{Set up a learndown Learnr application}
\usage{
learndownLearnrSetup(
  config,
  fingerprint,
  time.limit = 60,
  cap = "R Code",
  echo = FALSE,
  comment = NA,
  use.gradethis = TRUE,
  event.recorder = learndown::record_learnr,
  debug = Sys.getenv("LEARNDOWN_DEBUG", 0) != 0
)

learndownLearnrBanner(title, text, image, align = "left")
}
\arguments{
\item{config}{The \code{config()} command to use to get database info.}

\item{fingerprint}{The \code{fingerprint()} command to use to get user info.}

\item{time.limit}{The maximum time allowed to evaluate R code.}

\item{cap}{The caption for R code widgets.}

\item{echo}{Do we echo commands in R chunks?}

\item{comment}{The prefix added before each line of R chunk output.}

\item{use.gradethis}{Do we use {gradethis}?}

\item{event.recorder}{The function to use as event recorder. you should
probably not change the default value here.}

\item{debug}{Do we issue additions debugging informations?}

\item{title}{The Title for the banner.}

\item{text}{Text to print beneath the title.}

\item{image}{URL to an image to display in the banner.}

\item{align}{How is the image aligned: "left" (default), "right", "middle",
"top" or "bottom".}
}
\value{
Nothing. The function is used to setup the learnr environment.
}
\description{
This function eases the configuration of the learnr document to get user and
database info, record events, use grade this and parameterize learnr.
}
