% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learndown_shiny.R
\name{learndownShiny}
\alias{learndownShiny}
\alias{learndownShinyVersion}
\alias{submitAnswerButton}
\alias{quitButton}
\alias{submitQuitButtons}
\title{Create and manage learndown Shiny applications}
\usage{
learndownShiny(title, windowTitle = title)

learndownShinyVersion(version)

submitAnswerButton(
  inputId = "submit",
  label = "Submit",
  class = "btn-primary",
  ...
)

quitButton(inputId = "quit", label = "Quit", class = "btn-secondary", ...)

submitQuitButtons()
}
\arguments{
\item{title}{The title of the Shiny application}

\item{windowTitle}{The title of the window that holds the Shiny application,
by default, it is the same as \code{title}. If \code{title = NULL}, no title is added.}

\item{version}{The version number (in a string character format) of the Shiny
application to use in the events logger.}

\item{inputId}{The identifier of the button ("submit" or "quit").}

\item{label}{The button text ("Submit Answer" or "Quit").}

\item{class}{The bootstrap class of the button.}

\item{...}{Further arguments passed to \code{\link[shiny:actionButton]{shiny::actionButton()}}.}
}
\value{
The code to be inserted at the beginning of the Shiny application UI
for \code{\link[=learndownShiny]{learndownShiny()}} or where the buttons should be located for the other
function.
}
\description{
A learndown Shiny application is an application whose events (start, stop,
inputs, outputs, errors, result, quit) are recorded. It also provides a
'Submit Answer' and a 'Quit' buttons that manage to check the answer provided
by the user and to close the application cleanly.
}
\examples{
learndownShiny("My title")
}
\seealso{
\code{\link[=trackEvents]{trackEvents()}}, \code{\link[=record_shiny]{record_shiny()}}
}
