% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learnitdown_learnr.R
\name{learnitdownLearnrSetup}
\alias{learnitdownLearnrSetup}
\alias{learnitdownLearnrBanner}
\alias{learnitdownLearnrServer}
\title{Set up a learnitdown Learnr application}
\usage{
learnitdownLearnrSetup(
  config,
  sign_in,
  time.limit = 60,
  cap = "R Code",
  echo = FALSE,
  comment = NA,
  use.gradethis = TRUE,
  event.recorder = learnitdown::record_learnr,
  debug = Sys.getenv("LEARNITDOWN_DEBUG", 0) != 0
)

learnitdownLearnrBanner(
  title,
  text,
  image,
  align = "left",
  msg.nologin = "Anonymous user, no record!",
  msg.login = "Recording activated for ",
  msg.error = "Error recording activity! "
)

learnitdownLearnrServer(input, output, session)
}
\arguments{
\item{config}{The \code{config()} command to use to get database info.}

\item{sign_in}{The \code{sign_in()} command to use to get user info.}

\item{time.limit}{The maximum time allowed to evaluate R code.}

\item{cap}{The caption for R code widgets.}

\item{echo}{Do we echo commands in R chunks?}

\item{comment}{The prefix added before each line of R chunk output.}

\item{use.gradethis}{Do we use {gradethis}?}

\item{event.recorder}{The function to use as event recorder. you should
probably not change the default value here.}

\item{debug}{Do we issue additions debugging informations?}

\item{title}{The Title for the banner.}

\item{text}{Text to print beneath the title.}

\item{image}{URL to an image to display in the banner.}

\item{align}{How is the image aligned: "left" (default), "right", "middle",
"top" or "bottom".}

\item{msg.nologin}{The message to display if no user is logged in.}

\item{msg.login}{The message to display if a user is logged in (will be
followed by the login).}

\item{msg.error}{The message when an error during recording of activity in
the database occurs.}

\item{input}{The Shiny input.}

\item{output}{The Shiny output.}

\item{session}{The Shiny session.}
}
\value{
Nothing. The function is used to setup the learnr environment.
}
\description{
This function eases the configuration of the learnr document to get user and
database info, record events, use grade this and parameterize learnr.
}
