% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learnitdown_shiny.R
\name{learnitdownShiny}
\alias{learnitdownShiny}
\alias{learnitdownShinyVersion}
\alias{submitAnswerButton}
\alias{quitButton}
\alias{submitQuitButtons}
\title{Create and manage learnitdown Shiny applications}
\usage{
learnitdownShiny(title, windowTitle = title)

learnitdownShinyVersion(version)

submitAnswerButton(
  inputId = "learnitdown_submit_",
  label = "Submit",
  class = "btn-primary",
  ...
)

quitButton(
  inputId = "learnitdown_quit_",
  label = "Quit",
  class = "btn-secondary",
  ...
)

submitQuitButtons()
}
\arguments{
\item{title}{The title of the Shiny application}

\item{windowTitle}{The title of the window that holds the Shiny application,
by default, it is the same as \code{title}. If \code{title = NULL}, no title is added.}

\item{version}{The version number (in a string character format) of the Shiny
application to use in the events logger.}

\item{inputId}{The identifier of the button ("learnitdown_submit_" or
"learnitdown_quit_").}

\item{label}{The button text ("Submit" or "Quit").}

\item{class}{The bootstrap class of the button.}

\item{...}{Further arguments passed to \code{\link[shiny:actionButton]{shiny::actionButton()}}.}
}
\value{
The code to be inserted at the beginning of the Shiny application UI
for \code{\link[=learnitdownShiny]{learnitdownShiny()}} or where the buttons should be located for the other
function.
}
\description{
A learnitdown Shiny application is an application whose events (start, stop,
inputs, outputs, errors, result, quit) are recorded. It also provides a
'Submit Answer' and a 'Quit' buttons that manage to check the answer provided
by the user and to close the application cleanly.
}
\examples{
learnitdownShiny("My title")
}
\seealso{
\code{\link[=trackEvents]{trackEvents()}}, \code{\link[=record_shiny]{record_shiny()}}
}
