% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learnr.R
\name{learnr}
\alias{learnr}
\title{Insert a block for a learnr tutorial}
\usage{
learnr(
  id,
  title = NULL,
  package,
  toc = "",
  text = "Now let's make the exercises in the following tutorial:",
  toc.def = "Tutorial {id}",
  rstudio.url = "start_rstudio.html",
  connect.url = NULL,
  tuto.img = "images/list-tuto.png",
  tuto.link = "tutorial",
  icourse = "",
  institution = "",
  acad_year = "",
  term = "",
  set = ""
)
}
\arguments{
\item{id}{Identifier of the learnr tutorial (as \code{tutorial:id} in the YAML
section).}

\item{title}{The title of the tutorial (\code{title} in the YAML section). If
\code{NULL}, the \code{id} is used instead.}

\item{package}{Package where the learnr tutorial is defined.}

\item{toc}{Entry to use in the exercises table of content (\code{NULL} if no
entry, \code{""} for a default entry based on \verb{toc.def =}).}

\item{text}{The text to display in the learnr block.}

\item{toc.def}{Text for a default toc entry using \code{\link[=glue]{glue()}} syntax for
replacement, e.g., \code{{id}}.}

\item{rstudio.url}{The URL to open a page in RStudio server in the SciViews
box.}

\item{connect.url}{The URL to open a page in Posit Connect. If provided, it
is used instead of \code{rstudio.url}.}

\item{tuto.img}{The image to display in front of the toc entry}

\item{tuto.link}{The link when the image is clicked (sends to an help page
about learnr tutorials).}

\item{icourse}{The course identifier, e.g., \code{"MATH101"}.}

\item{institution}{The institution name, e.g., \code{"My University"}.}

\item{acad_year}{The academic year, e.g., \code{"2023-2024"}.}

\item{term}{The term, e.g., \code{"Q1"}.}

\item{set}{The set identifier, e.g., \code{"21M"} where 21 is the year and M is a
set identifier.}
}
\value{
The Markdown chunk to insert a learnr tutorial block in the document.
}
\description{
Insert Markdown text to link to a learnr tutorial.
}
