% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learndown_learnr.R
\name{learnr_user_info}
\alias{learnr_user_info}
\title{Get user information to properly record learnr events}
\usage{
learnr_user_info(
  session,
  input,
  output,
  path = Sys.getenv("LEARNDOWN_LOCAL_STORAGE", "~/.local/share/R"),
  debug = Sys.getenv("LEARNDOWN_DEBUG", 0) != 0
)
}
\arguments{
\item{session}{The current Shiny \code{session}.}

\item{input}{The Shiny \code{input} object.}

\item{output}{The Shiny \code{output} object.}

\item{path}{The path where a cached version of the user informations is
stored.}

\item{debug}{Do we debug recording of events using extra messages? By
default, it is the value of the environment variable \code{LEARNDOWN_DEBUG}, and
debugging is activated when that value is different to \code{0}.}
}
\value{
The code to be inserted in the server part of the learndown learnr
application in order to properly identify the user and record the events.
}
\description{
The server-side of a Shiny application to place in a learnr in order to
properly identify the user, and if login is provided, recording is done,
otherwise not.
}
\seealso{
\code{\link[=learndownShinyVersion]{learndownShinyVersion()}}
}
