% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obfuscate.R
\name{obfuscate}
\alias{obfuscate}
\alias{._}
\alias{obfuscate_logical}
\alias{ans}
\title{Obfuscate answers in learnr documents}
\usage{
obfuscate(object, password)

._(object)

obfuscate_logical(x, y, r = FALSE)

ans(correct, text, message = NULL)
}
\arguments{
\item{object}{An R object to obfuscate, or Obfuscation string to decrypt.}

\item{password}{The password to use (otherwise, the password defined by the
application is used)}

\item{x}{A logical value or an integer.}

\item{y}{Obfuscation parameter.}

\item{r}{Reset obfuscation (not intended for end-user)?}

\item{correct}{Is the answer is correct or not (either a logical value, or
an integer that obfuscates the results).}

\item{text}{Text of the option.}

\item{message}{Message displayed if the item is selected}
}
\value{
The encrypted or decrypted object.
}
\description{
Obfuscate answers in learnr documents
}
\examples{
obfuscate("test", "password")
}
