% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learndown_shiny.R
\name{read_shinylogs}
\alias{read_shinylogs}
\title{Read shinylogs log data and format them in a data.frame}
\usage{
read_shinylogs(file, version = "0", log.errors = TRUE, log.outputs = FALSE)
}
\arguments{
\item{file}{The path to the RDS file that contains the shinylogs log data.}

\item{version}{The version of the Shiny application. Version is not recorded
by shinylogs, so, we must provide it here.}

\item{log.errors}{Do we record errors too?}

\item{log.outputs}{Do we record outputs too (note that results are recorded
with inputs)?}
}
\value{
A data frame with the different events logged (one per line).
}
\description{
Learndown Shiny applications are a special kind of Shiny applications that
logs events and check results. It uses the \code{shinylogs} package to log Shiny
events and \code{\link[=read_shinylogs]{read_shinylogs()}} reads such logs and convert their data into a
format that is suitable to include, say in a MongoDB database.
}
\seealso{
\code{\link[=record_shiny]{record_shiny()}}, \code{\link[=trackEvents]{trackEvents()}}
}
