% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learndown_learnr.R
\name{record_learnr}
\alias{record_learnr}
\alias{user_name}
\alias{user_email}
\title{Record results of learnr exercises in a MongoDB database}
\usage{
record_learnr(tutorial_id, tutorial_version, user_id, event, data)

user_name(value)

user_email(value)
}
\arguments{
\item{tutorial_id}{The identifier of the tutorial.}

\item{tutorial_version}{The version of the tutorial.}

\item{user_id}{The user identifier for this learnr process.}

\item{event}{The event that triggers the record, like \code{exercise_submission}
or \code{question_submission}}

\item{data}{A JSON field with event-dependent data content. If \code{NULL}, only
a test to see if the database is responding is performed.}

\item{value}{The new value for user name or email (if not provided, the
current value is returned).}
}
\value{
Nothing. The function is used for its side-effects.
}
\description{
Record tutorial submissions in a MongoDB database. The
function is used by learndown learnr tutorials and is not for end-users.
}
\seealso{
\code{\link[=send_mail_learnr]{send_mail_learnr()}}
}
\concept{record events from the BioDataScience package}
\keyword{utilities}
