% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learnitdown_shiny.R
\name{record_shiny}
\alias{record_shiny}
\title{Record Shiny events in a MongoDB database}
\usage{
record_shiny(
  path,
  url,
  db,
  collection = "shiny",
  version = "0",
  log.errors = TRUE,
  log.outputs = FALSE,
  drop.dir = FALSE,
  debug = Sys.getenv("LEARNITDOWN_DEBUG", 0) != 0
)
}
\arguments{
\item{path}{The directory that contains shinylogs .rds files}

\item{url}{The mongodb url.}

\item{db}{The database name.}

\item{collection}{The name of the collection where to insert the documents.}

\item{version}{The version of the running Shiny application.}

\item{log.errors}{Do we record errors too?}

\item{log.outputs}{Do we record outputs too (note that results are recorded
with inputs)?}

\item{drop.dir}{If \code{TRUE} and path is empty at the end of the process, drop
the logs directory.}

\item{debug}{Do we debug the events recording by issuing extra messages? By
Default the value in the environment variable \code{LEARNITDOWN_DEBUG} is used and
debugging will be done if this value is different to \code{0}.}
}
\value{
\code{TRUE}if there where log files to export, \code{FALSE} otherwise.
}
\description{
Given a path that contains \code{shinylogs} events in .rds format, read these
events and transfer them into a MongoDB database.
}
\seealso{
\code{\link[=read_shinylogs]{read_shinylogs()}}, \code{\link[=trackEvents]{trackEvents()}}
}
