% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learndown_learnr.R
\name{run}
\alias{run}
\title{Run learndown learnr tutorials from a package}
\usage{
run(tutorial, package, repos = NULL, ..., update = ask, ask = interactive())
}
\arguments{
\item{tutorial}{The name of the tutorial to use. If not provided, a list of
available tutorials is displayed.}

\item{package}{The package from where to run the tutorial.}

\item{repos}{The GitHub repository where the package is developed (for
updates), use \code{NULL} to prevent any updates.}

\item{...}{Further arguments passed to \code{\link[=run_tutorial]{run_tutorial()}}}

\item{update}{Do we check for an updated version first, and if it is found,
update the package automatically?}

\item{ask}{In case \code{tutorial} is not provided, do we ask to select in a list?}
}
\value{
If \code{tutorial} is not provided, in interactive mode with \code{ask = TRUE},
you have to select one in a list, and in non interactive mode, or
\code{ask = FALSE}, it returns the list of all available tutorials.
}
\description{
Start the learnr R engine in the current R session with the
selected tutorial.
}
\examples{
\dontrun{
#' # To start from a list of available tutorials:
run(package = "my_package")
run("my_tutorial", package = "my_package")
}
}
\seealso{
\code{\link[=run_tutorial]{run_tutorial()}}
}
\concept{run interactive learnr documents from the BioDataScience package}
\keyword{utilities}
