% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learndown_learnr.R
\name{send_mail_learnr}
\alias{send_mail_learnr}
\title{Send your learnr submissions by email}
\usage{
send_mail_learnr(
  address,
  subject = "Learnr activity",
  file = Sys.getenv("LOCAL_STORAGE")
)
}
\arguments{
\item{address}{The mail address to send the data to.}

\item{subject}{The title of the mail.}

\item{file}{The file that contains your learnr activity information.}
}
\value{
The data are returned invisibly.
}
\description{
Your submissions are send to a central database. However, in
case that database is not accessible, the data is stored locally. This
function uses your plain email to send your records. Note that, once the
email is created, the local version of your records is reset. So, if you
finally decide to NOT send the email, these records are lost (in this case,
call your teachers to recover them, if you have to.)
}
\details{
In case the MongoDB cannot be reach, learnr events are stored in a local
file. This function allows to submits its content through email as an
alternate way to collect learnr activity.
}
\examples{
\dontrun{
send_mail_learnr("me\@mymail.org")
}
}
\seealso{
\code{\link[=record_learnr]{record_learnr()}}
}
\concept{run interactive learnr documents from the BioDataScience package}
\keyword{utilities}
