% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_ex_toc.R
\name{show_ex_toc}
\alias{show_ex_toc}
\alias{clean_ex_toc}
\title{Insert a table of content for the exercises at the end of a bookdown chapter}
\usage{
show_ex_toc(id, header = "", clear.it = TRUE, finalize = FALSE)

clean_ex_toc()
}
\arguments{
\item{id}{An identifier for this table of content.}

\item{header}{A Markdown text to place as header of the exercises toc.}

\item{clear.it}{Do we clear the toc list (\code{TRUE} by default)?}

\item{finalize}{If \code{TRUE}, we also compile the \code{assignments.csv} and
\code{apps.csv} files with all assignments and all apps.}
}
\value{
The Markdown chunk with the exercises toc.
}
\description{
For the various exercise types (h5p, shiny apps, learnrs & GitHub
assignments/challenges) we add toc entries with \code{\link[=h5p]{h5p()}}, \code{\link[=launch_shiny]{launch_shiny()}},
\code{\link[=learnr]{learnr()}}, \code{\link[=assignment]{assignment()}}, and \code{\link[=challenge]{challenge()}} respectively. This function
creates the table of content for all these exercises.
}
