% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learndown_shiny.R
\name{trackEvents}
\alias{trackEvents}
\alias{trackSubmit}
\alias{trackQuit}
\alias{check_shiny_solution}
\title{Track events, the submit or the quit buttons}
\usage{
trackEvents(
  session,
  input,
  output,
  sign_in.fun = NULL,
  url = Sys.getenv("MONGO_URL"),
  url.server = Sys.getenv("MONGO_URL_SERVER"),
  db = Sys.getenv("MONGO_BASE"),
  user = Sys.getenv("MONGO_USER"),
  password = Sys.getenv("MONGO_PASSWORD"),
  version = getOption("learndown.shiny.version"),
  path = Sys.getenv("LEARNDOWN_LOCAL_STORAGE", "shiny_logs"),
  log.errors = TRUE,
  log.outputs = FALSE,
  drop.dir = TRUE,
  debug = Sys.getenv("LEARNDOWN_DEBUG", 0) != 0
)

trackSubmit(
  session,
  input,
  output,
  solution = NULL,
  max_score = NULL,
  comment = "",
  message.success = "Correct",
  message.error = "Incorrect",
  score.txt = "Score",
  check.solution = check_shiny_solution
)

trackQuit(session, input, output, delay = 60)

check_shiny_solution(answer, solution)
}
\arguments{
\item{session}{The current Shiny \code{session}.}

\item{input}{The Shiny \code{input} object.}

\item{output}{The Shiny \code{output} object.}

\item{sign_in.fun}{The function that can get user info from the
\code{\link[=sign_in]{sign_in()}}, or \code{NULL} by default to disable using local user data.}

\item{url}{The URL to reach the MongoDB database. By default, it is read from
the \code{MONGO_URL} environment variable.}

\item{url.server}{The URL to reach the MongoDB database. By default, it is
read from the \code{MONGO_URL_SERVER} environment variable.}

\item{db}{The database to populate in the MongoDB database. By default, it is
read from the \code{MONGO_BASE} environment variable.}

\item{user}{The user login to the MongoDB database. By default, it is
read from the \code{MONGO_USER} environment variable.}

\item{password}{The password to access the MongoDB database. By default, it
is read from the \code{MONGO_PASSWORD} environment variable.}

\item{version}{The version of the current Shiny application. By default, it
is the \code{learndown.shiny.version} option, as set by \code{\link[=learndownShinyVersion]{learndownShinyVersion()}}.}

\item{path}{The path where the temporary \code{shinylogs} log files are stored. By
default, it is set to the \code{LEARNDOWN_LOCAL_STORAGE} environment variable,
or to \code{shiny_logs} subdirectory of the application if not defined. If that
directory is not writable, a temporary directory is used instead.}

\item{log.errors}{Do we log error events (yes by default)?}

\item{log.outputs}{Do we log output events (no by default)?}

\item{drop.dir}{Do we erase the directory indicated by \verb{path =} if it is
empty at the end of the process (yes by default).}

\item{debug}{Do we debug recording of events using extra messages? By
default, it is the value of the environment variable \code{LEARNDOWN_DEBUG}, and
debugging is activated when that value is different to \code{0}.}

\item{solution}{The correct solution as a named list. Names are the
application inputs to check and their values are the correct values. The
current state of these inputs will be compared against the solution, and the
\code{message.success} or \code{message.error} is displayed accordingly. Also a result
event is triggered with the \code{correct} field set to \code{TRUE} or \code{FALSE}
accordingly. If \code{solution = NULL}, then the answer is considered to be always
correct (use this if you just want to indicate that the user's answer is
recorded in the database). For numeric values, use `c(min = x, max = y) to
check if values are within a range.}

\item{max_score}{The highest score value that could be attributed if the
exercise is correctly done. By default, it is \code{NULL} meaning that the higher
score would be equal to the number of items to check in \code{solution}.}

\item{comment}{A string with a comment to append to the value of the result
event.}

\item{message.success}{The message to display is the answer is correct
("Correct" by default).}

\item{message.error}{The message to display if the answer is wrong
("Incorrect" by default).}

\item{score.txt}{The word to use for "score" ("Score" by default).}

\item{check.solution}{The function to use to check if solution provided by
a Shiny application is correct or not. By default, it is
\code{check_shiny_solution()}.}

\item{delay}{The time to wait before we close the Shiny application in sec
(60 sec by default). If \code{delay = -1}, the application is \strong{not} closed, only
the session is closed when the user clicks on the quit button.}

\item{answer}{A named list of the answer data to check against solution. This
object is provided by \code{trackSubmit()}.}
}
\value{
The code to be inserted in the server part of the learndown Shiny
application in order to properly identify the user and record the events.
}
\description{
These functions provide the required code to be inserted in the server part
of a Shiny application to track events (start, stop, inputs, outputs, errors,
result or quit), to check the answer when the user clicks on the submit
button and to cleanly close the application when the user clicks on the quit
button.
}
\seealso{
\code{\link[=learndownShinyVersion]{learndownShinyVersion()}}, \code{\link[=sign_in]{sign_in()}}
}
