% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webshot_shiny.R
\name{webshot_shiny}
\alias{webshot_shiny}
\title{Create the screenshot image of a Shiny application with a click icon}
\usage{
webshot_shiny(
  url,
  app = basename(url),
  imgdir = "images/shinyapps",
  img = paste0(imgdir, "/", app, ".png"),
  width = 780,
  height = 500,
  offsetx = 30,
  offsety = 30,
  delay = 10,
  ...
)
}
\arguments{
\item{url}{The URL to launch the Shiny app. If both \verb{app =} and \verb{baseurl =}}

\item{app}{The name of the Shiny application.
are provided, you don't need to specify it.}

\item{imgdir}{The directory without trailing "/" where images relative
to Shiny applications are stored. By default, it is relative to current
directory, in \code{images/shinyapps} subdirectories.}

\item{img}{The path to the image that is created. Not needed if \verb{app =} and
\verb{imgdir =} are provided.}

\item{width}{The requested weight of the screenshot (it may differ if the
Shiny application defines other (limit) values.}

\item{height}{The requested height of the screenshot (idem).}

\item{offsetx}{The offset from left where to place the click icon in pixels.}

\item{offsety}{The offset to bottom where to place the click icon in pixels.}

\item{delay}{Time to wait (in sec) after the Shiny application has started
and before the screenshot is taken. If the screenshot does not contain the
complete application UI, try increase this value.}

\item{...}{Further arguments passed to \code{\link[=launch_shiny]{launch_shiny()}} but not used here.}
}
\value{
The path to the created image, invisibly.
}
\description{
The \code{\link[=webshot_shiny]{webshot_shiny()}} function is designed to create the screenshot image of
a Shiny application, with an icon suggesting to click on it for launching
the application. The image created can then be used by \code{\link[=launch_shiny]{launch_shiny()}} in
a bookdown to differ the start of Shiny application, while displaying useful
information to the user (how the Shiny application would look like if it was
started).
}
\examples{
\dontrun{
# We wait 10 sec to make sure it is loaded when the screenshot is taken
(webshot_shiny("https://phgrosjean.shinyapps.io/histogram/", delay = 10))
 # Now, look at this image. You can use it with launch_shiny()
}
}
\seealso{
\code{\link[=launch_shiny]{launch_shiny()}}
}
