% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config}
\alias{config}
\alias{encrypt}
\alias{decrypt}
\title{Configure the R environment for the course (including database information)}
\usage{
config(url, password, cache = ".learndown_config")

encrypt(object, password)

decrypt(object, password)
}
\arguments{
\item{url}{The URL of the encrypted file that contains the configuration
information.}

\item{password}{The password to decrypt the data.}

\item{cache}{The path to the file to use to store a cached version of these
data. Access to the database will be checked, and if it fails, the
configuration data are refreshed from the URL.}

\item{object}{An object to be encrypted.}
}
\value{
Invisibly returns \code{TRUE} if success, or \code{FALSE} otherwise for
\code{\link[=config]{config()}}. The encrypted/decrypted object for \code{\link[=encrypt]{encrypt()}} and \code{\link[=decrypt]{decrypt()}}.
}
\description{
Call this function every time you need to get environment variables set, like
the URL, user and password of the MongoDB database used by the course.
}
