% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_shiny.R
\name{launch_shiny}
\alias{launch_shiny}
\title{Launch Shiny application by clicking on its screenshot.}
\usage{
launch_shiny(
  url,
  app = basename(url),
  imgdir = "images/shinyapps",
  img = paste0(imgdir, "/", app, ".png"),
  createimg = TRUE,
  width = 780,
  height = 500,
  fun = NULL,
  alt1 = "*Click to start the Shiny application.*",
  alt2 = paste0("*Click to start",
    "or [run `{run.cmd}`]({run.url}{run.arg}){{target=\\"_blank\\"}}.*"),
  toc = "",
  toc.def = "Shiny application {app}",
  run.url = "start_rstudio.html?runrcode=",
  run.cmd = glue("{fun}(\\"{app}\\")"),
  run.arg = URLencode(run.cmd, reserved = TRUE),
  app.img = "images/list-app.png",
  app.link = "shiny_app",
  ...
)
}
\arguments{
\item{url}{The URL of the Shiny application. If both \verb{app =} and \verb{baseurl =}
are provided, you don't need to specify it.}

\item{app}{Name of the shiny application (cannot be duplicated on a page).}

\item{imgdir}{The directory without trailing "/" where images relative
to Shiny applications are stored. By default, it is relative to current
directory, in \code{images/shinyapps} subdirectories.}

\item{img}{The relative or absolute path to the image with a screenshot of
the Shiny application, as produced by \code{\link[=webshot_shiny]{webshot_shiny()}}.}

\item{createimg}{If the app image (\code{img}) is not found, and there is no
default image in \verb{imgdir =}, do we put it there (yes be default)?}

\item{width}{The width of the image and iframe for the app.}

\item{height}{The height of image and iframe.}

\item{fun}{The function to run as alternative to start the Shiny application
locally. It is better to fully specify it (\verb{package::function}), and it
should take one argument which is the application name in \verb{app =}.}

\item{alt1}{Alternate text to display at the bottom of the screenshot when
nothing is provided for \verb{fun =}. If
\code{NULL}, nothing is displayed below the screenshot.}

\item{alt2}{Alternate text to display at the bottom of the screenshot in case
\verb{fun =} is provided.}

\item{toc}{Entry to use in the exercises table of content (\code{NULL} if no
entry, \code{""} for a default entry based on \verb{toc.def =}).}

\item{toc.def}{Text for a default toc entry using \code{\link[=glue]{glue()}} syntax for
replacement, e.g., \code{{app}}.}

\item{run.url}{The URL to use to start the Shiny application in RStudio
server in the SciViews Box. It should generally end with \verb{?runrcode=}, and
the R code to execute will be appended to it from \verb{run.arg =}.}

\item{run.cmd}{The command to use to launch the Shiny application in RStudio.}

\item{run.arg}{The URL encoded version of \verb{run.cmd =}.}

\item{app.img}{The image to display in front of the toc entry}

\item{app.link}{The link when the image is clicked (sends to an help page
about Shiny applications).}

\item{...}{Not used here, but it allows to add more arguments used by the
screenshot addin, like \verb{delay =}, \verb{offsetx =} or \verb{offsety =}, see
\code{\link[=webshot_shiny]{webshot_shiny()}}.}
}
\value{
The HTML content that creates the image and the iframe. The function
must be called from within an R inline expression or from an R chunk with
\code{results='asis'} in an HTML-rendered version of the R Markdown document to
get the correct result.
}
\description{
Shiny applications can be embedded in certain R Markdown documents. However,
the application is automatically loaded at the same time as the main page,
and that may not be the desired behavior. With \code{\link[=launch_shiny]{launch_shiny()}}, you display
just a screenshot of the Shiny application in the page. The user has to click
on it to actually launch the application (that replaces the screenshot on the
page).
}
\examples{
# TODO...
}
\seealso{
\code{\link[=webshot_shiny]{webshot_shiny()}}
}
